New filter to embed a Sketchcast sketch (http://sketchcast.com/) by using the document code of the Sketchcast sketch. This takes the format:
[SKETCHCAST: id=CODE]

For example,

	[SKETCHCAST: id=Woo8w3l]

The CODE is the part of the URL after /view/ (http://sketchcast.com/view/Woo8w3l), so in the example shown, the id is Woo8w3l.

You can also add optional width and height parameters in the format

	[SKETCHCAST: id=Woo8w3l, width=WIDTH, height=HEIGHT]

where WIDTH and HEIGHT are given in pixels.